+++ node/src/parachain/mod.rs
@@ -10,7 +10,7 @@ mod shell_upgrade;
 pub mod service;
 
 /// Parachain specs.
-pub mod dev_chain_spec;
+pub mod agung_chain_spec;
 pub mod krest_chain_spec;
 pub mod peaq_chain_spec;
